Moved common/baseDBMethods to routes/baseDBMethods
